BE/requirements.txt
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.7
astroid==2.4.1
atomicwrites==1.4.0
attrs==19.3.0
bcrypt==3.1.7
billiard==3.6.3.0
cached-property==1.5.1
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
Django==3.0.6
django-celery==3.3.1
django-celery-results==1.2.1
django-cors-headers==3.3.0
docker==4.2.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
gevent==20.5.1
greenlet==0.4.15
idna==2.9
isort==4.3.21
jsonschema==3.2.0
kombu==4.6.8
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
paho-mqtt==1.5.0
paramiko==2.7.1
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pylint==2.5.2
PyNaCl==1.3.0
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.16.0
pytest==5.4.3
pytz==2020.1
pywin32==227
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.17
sqlparse==0.3.1
texttable==1.6.2
toml==0.10.1
urllib3==1.25.9
vine==1.3.0
wcwidth==0.2.5
websocket-client==0.57.0
wrapt==1.12.1
zope.event==4.4
zope.interface==5.1.0